# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AutoSSL 证书管理API文档",
        "title": "AutoSSL API",
        "termsOfService": "http://example.com/terms",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "http://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/apiv1",
    "paths": {
...
        "/workflow/workflows": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取当前用户的工作流列表，支持分页和筛选",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作流管理"
                ],
                "summary": "获取工作流列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数，默认10",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "开始日期，格式：2006-01-02",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期，格式：2006-01-02",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "状态筛选，可选值：pending|running|completed|failed|cancelled",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "域名筛选，支持模糊匹配",
                        "name": "domain",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "工作流列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/ent.Workflow"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "获取工作流列表失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/workflow/workflows/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据工作流ID获取工作流的详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作流管理"
                ],
                "summary": "获取工作流详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "工作流ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "工作流详情",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/ent.Workflow"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "无效的工作流ID",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "获取工作流失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/workflow/workflows/{id}/cancel": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "取消正在运行的工作流",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作流管理"
                ],
                "summary": "取消工作流",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "工作流ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "工作流已取消",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "无效的工作流ID",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "取消工作流失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/workflow/workflows/{id}/retry": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "重试失败的工作流",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作流管理"
                ],
                "summary": "重试工作流",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "工作流ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "工作流已重新启动",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "无效的工作流ID",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "重试工作流失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/workflow/workflows/{id}/start": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据工作流ID启动工作流执行",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作流管理"
                ],
                "summary": "启动工作流",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "工作流ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "工作流已成功启动",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "无效的工作流ID",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "启动工作流失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/workflow/workflows/{id}/status": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "更新指定工作流的状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作流管理"
                ],
                "summary": "更新工作流状态",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "工作流ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "状态更新请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "工作流状态已更新",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "无效的工作流ID或请求参数",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "更新工作流状态失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "certificate.Status": {
            "type": "string",
            "enum": [
                "pending",
                "pending",
                "processing",
                "issued",
                "deploying",
                "completed",
                "failed"
            ],
            "x-enum-varnames": [
                "DefaultStatus",
                "StatusPending",
                "StatusProcessing",
                "StatusIssued",
                "StatusDeploying",
                "StatusCompleted",
                "StatusFailed"
            ]
        },
        "certificate.ValidationType": {
            "type": "string",
            "enum": [
                "dns",
                "http"
            ],
            "x-enum-varnames": [
                "ValidationTypeDNS",
                "ValidationTypeHTTP"
            ]
        },
...
        "ent.Certificate": {
            "type": "object",
            "properties": {
                "access_ids": {
                    "description": "访问配置ID列表",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "acme_account_email": {
                    "description": "ACME账号邮箱",
                    "type": "string"
                },
                "created_at": {
                    "description": "CreatedAt holds the value of the \"created_at\" field.",
                    "type": "string"
                },
                "dns_provider_name": {
                    "description": "DNS提供商名称",
                    "type": "string"
                },
                "domains": {
                    "description": "证书域名列表",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the CertificateQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.CertificateEdges"
                        }
                    ]
                },
                "error_message": {
                    "description": "错误信息",
                    "type": "string"
                },
                "expired_at": {
                    "description": "过期时间",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "issued_at": {
                    "description": "签发时间",
                    "type": "string"
                },
                "issued_by": {
                    "description": "签发机构",
                    "type": "string"
                },
                "renewal_days_before": {
                    "description": "距离过期前多少天开始续期",
                    "type": "integer"
                },
                "renewal_enabled": {
                    "description": "是否启用自动续期",
                    "type": "boolean"
                },
                "server_group_id": {
                    "description": "服务器组ID",
                    "type": "integer"
                },
                "status": {
                    "description": "Status holds the value of the \"status\" field.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/certificate.Status"
                        }
                    ]
                },
                "updated_at": {
                    "description": "UpdatedAt holds the value of the \"updated_at\" field.",
                    "type": "string"
                },
                "validation_type": {
                    "description": "验证类型",
                    "allOf": [
                        {
                            "$ref": "#/definitions/certificate.ValidationType"
                        }
                    ]
                }
            }
        },
...
        "ent.CertificateEdges": {
            "type": "object",
            "properties": {
                "acme_account": {
                    "description": "AcmeAccount holds the value of the acme_account edge.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.ACMEAccount"
                        }
                    ]
                },
                "deployments": {
                    "description": "Deployments holds the value of the deployments edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.CertificateDeployment"
                    }
                },
                "dns_provider": {
                    "description": "DNSProvider holds the value of the dns_provider edge.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.DNSProvider"
                        }
                    ]
                },
                "user": {
                    "description": "User holds the value of the user edge.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.User"
                        }
                    ]
                },
                "workflow_steps": {
                    "description": "WorkflowSteps holds the value of the workflow_steps edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.WorkflowStep"
                    }
                },
                "workflows": {
                    "description": "关联的工作流",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Workflow"
                    }
                }
            }
        },
...
        "ent.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "CreatedAt holds the value of the \"created_at\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the UserQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.UserEdges"
                        }
                    ]
                },
                "email": {
                    "description": "登录邮箱",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "is_active": {
                    "description": "账户状态",
                    "type": "boolean"
                },
                "is_admin": {
                    "description": "管理员标识",
                    "type": "boolean"
                },
                "updated_at": {
                    "description": "UpdatedAt holds the value of the \"updated_at\" field.",
                    "type": "string"
                }
            }
        },
        "ent.UserEdges": {
            "type": "object",
            "properties": {
                "access_configs": {
                    "description": "访问配置",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Access"
                    }
                },
                "acme_accounts": {
                    "description": "用户的ACME账户",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.ACMEAccount"
                    }
                },
                "certificates": {
                    "description": "用户证书",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Certificate"
                    }
                },
                "dns_providers": {
                    "description": "DNS提供商配置",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.DNSProvider"
                    }
                },
                "notification_configs": {
                    "description": "通知配置",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.NotificationConfig"
                    }
                },
                "orders": {
                    "description": "用户订单",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Order"
                    }
                },
                "roles": {
                    "description": "用户角色",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Role"
                    }
                },
                "server_groups": {
                    "description": "服务器组",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.ServerGroup"
                    }
                },
                "workflow_steps": {
                    "description": "工作流步骤",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.WorkflowStep"
                    }
                },
                "workflows": {
                    "description": "用户工作流",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Workflow"
                    }
                }
            }
        },
        "ent.Workflow": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "description": "完成时间",
                    "type": "string"
                },
                "created_at": {
                    "description": "CreatedAt holds the value of the \"created_at\" field.",
                    "type": "string"
                },
                "current_step": {
                    "description": "当前工作流步骤",
                    "type": "integer"
                },
                "domain": {
                    "description": "域名",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the WorkflowQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.WorkflowEdges"
                        }
                    ]
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "status": {
                    "description": "工作流状态: pending, processing, issued, deploying, completed, failed, cancelled",
                    "type": "string"
                },
                "total_steps": {
                    "description": "总工作流步骤数",
                    "type": "integer"
                },
                "updated_at": {
                    "description": "UpdatedAt holds the value of the \"updated_at\" field.",
                    "type": "string"
                }
            }
        },
        "ent.WorkflowEdges": {
            "type": "object",
            "properties": {
                "certificate": {
                    "description": "Certificate holds the value of the certificate edge.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.Certificate"
                        }
                    ]
                },
                "steps": {
                    "description": "Steps holds the value of the steps edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.WorkflowStep"
                    }
                },
                "user": {
                    "description": "User holds the value of the user edge.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.User"
                        }
                    ]
                }
            }
        },
        "ent.WorkflowStep": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "description": "完成时间",
                    "type": "string"
                },
                "created_at": {
                    "description": "CreatedAt holds the value of the \"created_at\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the WorkflowStepQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.WorkflowStepEdges"
                        }
                    ]
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "log": {
                    "description": "步骤日志",
                    "type": "string"
                },
                "name": {
                    "description": "步骤名称",
                    "type": "string"
                },
                "started_at": {
                    "description": "开始时间",
                    "type": "string"
                },
                "status": {
                    "description": "步骤状态",
                    "allOf": [
                        {
                            "$ref": "#/definitions/workflowstep.Status"
                        }
                    ]
                },
                "step_number": {
                    "description": "步骤编号",
                    "type": "integer"
                }
            }
        },
        "ent.WorkflowStepEdges": {
            "type": "object",
            "properties": {
                "certificate": {
                    "description": "Certificate holds the value of the certificate edge.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.Certificate"
                        }
                    ]
                },
                "user": {
                    "description": "所属用户",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.User"
                    }
                },
                "workflow": {
                    "description": "所属工作流",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.Workflow"
                        }
                    ]
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "workflowstep.Status": {
            "type": "string",
            "enum": [
                "pending",
                "pending",
                "processing",
                "issued",
                "deploying",
                "completed",
                "failed",
                "cancelled"
            ],
            "x-enum-varnames": [
                "DefaultStatus",
                "StatusPending",
                "StatusProcessing",
                "StatusIssued",
                "StatusDeploying",
                "StatusCompleted",
                "StatusFailed",
                "StatusCancelled"
            ]
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}